+++ CHANGELOG.md
@@ -71,6 +71,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Added support for `Material` background textures with INI property `BGTextureFile`.
 
+- New `MovableObject` INI and Lua (R/W) property `SimUpdatesBetweenScriptedUpdates`, that lets `MovableObject`s run their Lua update function less frequently, for performance benefits.
+
 </details>
 
 <details><summary><b>Changed</b></summary>